feat: separate modal and form
